requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
click==7.1.2
croniter==1.0.11
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
python-dotenv==0.19.2
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1
